refactor: Game 삭제 및 역할 분할, 클래스명 변경
